Restructuring solution, will continue working chios method but need to make code more modular. Also working on a general linear system solver.
