+++ SystemLinearEquations/LinearSystemAlgorithms/CramersRule.cs
@@ -1,0 +1,19 @@
+﻿using Maths.LinearAlgebra;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SystemLinearEquations.LinearSystemAlgorithms
+{
+    // niave implementation
+    // just for demonstration purposes
+    public class CramersRule : ILinearSystemMethod
+    {
+        public double[] Solve(Matrix A, double[] b)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
