+++ SystemLinearEquations/LinearSystemAlgorithms/ILinearSystemMethod.cs
@@ -1,0 +1,9 @@
+﻿using Maths.LinearAlgebra;
+
+namespace SystemLinearEquations.LinearSystemAlgorithms
+{
+    public interface ILinearSystemMethod
+    {
+        public double[] Solve(Matrix A, double[] b);
+    }
+}
