+++ SystemLinearEquations/SystemLinearEquationsTests/LinearSystemSolverUnitTest.cs
@@ -1,0 +1,12 @@
+﻿
+namespace SystemLinearEquationsTests
+{
+    public static class LinearSystemSolverUnitTest
+    {
+        [Fact]
+        public static void TestSolveSystem()
+        {
+
+        }
+    }
+}
